【修复】商品类别详细，点击崩溃bug

+++ app/src/main/java/com/tts/starsky/phonesweepcode/controller/ContentManager.java
@@ -15,7 +15,7 @@ public class ContentManager {
         this.context = context;
     }
 
-    public List<GoodsInfo> getGoodsInfoList(int type_concrete_id) {
+    public List<GoodsInfo> getGoodsInfoList(Long type_concrete_id) {
 
         /*DatabaseOP databaseOP = new DatabaseOP(context);
         List<CommodityInfo> commodityInfoList = databaseOP.getGoodsInfoList(type_concrete_id);*/
